test/requirements.txt
requests==2.23.0
pytest==5.4.1
locustio==0.14.5